# adapt/docs/openapi.json
{
 "components": {
  "schemas": {
...
   "ArchiveOut": {
    "properties": {
     "artifacts": {
      "items": {
       "$ref": "#/components/schemas/ArchiveArtifact"
      },
      "title": "Artifacts",
      "type": "array"
     },
     "code_sha": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Code Sha"
     },
     "decision_hash": {
      "title": "Decision Hash",
      "type": "string"
     },
     "decision_id": {
      "title": "Decision Id",
      "type": "string"
     },
     "environment_fingerprint": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Environment Fingerprint"
     },
     "lock_hash": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lock Hash"
     },
     "note": {
      "title": "Note",
      "type": "string"
     },
     "seed": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seed"
     },
     "snapshot_id": {
      "title": "Snapshot Id",
      "type": "string"
     },
     "status": {
      "enum": [
       "AVAILABLE",
       "UNAVAILABLE"
      ],
      "title": "Status",
      "type": "string"
     },
     "steps": {
      "items": {
       "$ref": "#/components/schemas/ArchiveStep"
      },
      "title": "Steps",
      "type": "array"
     }
    },
    "required": [
     "decision_id",
     "decision_hash",
     "snapshot_id",
     "environment_fingerprint",
     "code_sha",
     "lock_hash",
     "seed",
     "status",
     "note",
     "artifacts",
     "steps"
    ],
    "title": "ArchiveOut",
    "type": "object"
   },
...
   "CopilotBody": {
    "properties": {
     "message": {
      "maxLength": 2000,
      "minLength": 1,
      "title": "Message",
      "type": "string"
     }
    },
    "required": [
     "message"
    ],
    "title": "CopilotBody",
    "type": "object"
   },
...
   "CreativeScore": {
    "properties": {
     "explanation": {
      "title": "Explanation",
      "type": "string"
     },
     "score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Score"
     },
     "status": {
      "enum": [
       "AVAILABLE",
       "NOT_ESTIMABLE"
      ],
      "title": "Status",
      "type": "string"
     }
    },
    "required": [
     "status",
     "score",
     "explanation"
    ],
    "title": "CreativeScore",
    "type": "object"
   },
   "Curve": {
    "properties": {
     "budget_id": {
      "title": "Budget Id",
      "type": "string"
     },
     "label": {
      "title": "Label",
      "type": "string"
     },
     "points": {
      "items": {
       "$ref": "#/components/schemas/CurvePoint"
      },
      "title": "Points",
      "type": "array"
     },
     "reason": {
      "title": "Reason",
      "type": "string"
     },
     "unit": {
      "title": "Unit",
      "type": "string"
     }
    },
    "required": [
     "budget_id",
     "label",
     "unit",
     "points",
     "reason"
    ],
    "title": "Curve",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "ReplayOut": {
    "properties": {
     "actual_hash": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Actual Hash"
     },
     "expected_hash": {
      "title": "Expected Hash",
      "type": "string"
     },
     "message": {
      "title": "Message",
      "type": "string"
     },
     "status": {
      "enum": [
       "VERIFIED",
       "MISMATCH",
       "UNAVAILABLE"
      ],
      "title": "Status",
      "type": "string"
     }
    },
    "required": [
     "status",
     "expected_hash",
     "actual_hash",
     "message"
    ],
    "title": "ReplayOut",
    "type": "object"
   },
...
   "SimulateBody": {
    "properties": {
     "decision_hash": {
      "title": "Decision Hash",
      "type": "string"
     }
    },
    "required": [
     "decision_hash"
    ],
    "title": "SimulateBody",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "ADAPT API",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/copilot/chat": {
   "post": {
    "description": "LLM offline (Stage 1): a deterministic template answer from stored engine state, streamed as the agreed SSE\ncontract (answer + done events), with citations to app routes only.",
    "operationId": "copilot_chat_api_v1_copilot_chat_post",
    "parameters": [
     {
      "in": "header",
      "name": "X-Request-ID",
      "required": true,
      "schema": {
       "title": "X-Request-Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": true,
      "schema": {
       "title": "Idempotency-Key",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CopilotBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Copilot Chat",
    "tags": [
     "insights"
    ]
   }
  },
  "/api/v1/copilot/sql": {
   "post": {
    "operationId": "copilot_sql_api_v1_copilot_sql_post",
    "parameters": [
     {
      "in": "header",
      "name": "X-Request-ID",
      "required": true,
      "schema": {
       "title": "X-Request-Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": true,
      "schema": {
       "title": "Idempotency-Key",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Copilot Sql",
    "tags": [
     "insights"
    ]
   }
  },
...
  "/api/v1/creatives/score": {
   "post": {
    "operationId": "creative_score_api_v1_creatives_score_post",
    "parameters": [
     {
      "in": "header",
      "name": "X-Request-ID",
      "required": true,
      "schema": {
       "title": "X-Request-Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": true,
      "schema": {
       "title": "Idempotency-Key",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CreativeScore"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Creative Score",
    "tags": [
     "insights"
    ]
   }
  },
  "/api/v1/curves/{budget_id}": {
   "get": {
    "operationId": "curve_api_v1_curves__budget_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "budget_id",
      "required": true,
      "schema": {
       "title": "Budget Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Curve"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Curve",
    "tags": [
     "insights"
    ]
   }
  },
...
  "/api/v1/decisions/{decision_id}/archive": {
   "get": {
    "operationId": "archive_api_v1_decisions__decision_id__archive_get",
    "parameters": [
     {
      "in": "path",
      "name": "decision_id",
      "required": true,
      "schema": {
       "title": "Decision Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ArchiveOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Archive",
    "tags": [
     "insights"
    ]
   }
  },
...
  "/api/v1/decisions/{decision_id}/replay": {
   "get": {
    "description": "Replay re-runs economics + optimizer + policy from the snapshot (seconds to tens of seconds), so it runs in a\nbackground thread and is cached per decision; until it finishes the answer is UNAVAILABLE (in progress).",
    "operationId": "replay_api_v1_decisions__decision_id__replay_get",
    "parameters": [
     {
      "in": "path",
      "name": "decision_id",
      "required": true,
      "schema": {
       "title": "Decision Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReplayOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Replay",
    "tags": [
     "insights"
    ]
   }
  },
  "/api/v1/decisions/{decision_id}/simulate": {
   "post": {
    "operationId": "simulate_api_v1_decisions__decision_id__simulate_post",
    "parameters": [
     {
      "in": "path",
      "name": "decision_id",
      "required": true,
      "schema": {
       "title": "Decision Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "X-Request-ID",
      "required": true,
      "schema": {
       "title": "X-Request-Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": true,
      "schema": {
       "title": "Idempotency-Key",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SimulateBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Comparison"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Simulate",
    "tags": [
     "insights"
    ]
   }
  },
  "/api/v1/decisions/{decision_id}/snapshot": {
   "get": {
    "operationId": "snapshot_api_v1_decisions__decision_id__snapshot_get",
    "parameters": [
     {
      "in": "path",
      "name": "decision_id",
      "required": true,
      "schema": {
       "title": "Decision Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SnapshotOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Snapshot",
    "tags": [
     "insights"
    ]
   }
  },
  "/api/v1/decisions/{decision_id}/timeline": {
   "get": {
    "operationId": "timeline_api_v1_decisions__decision_id__timeline_get",
    "parameters": [
     {
      "in": "path",
      "name": "decision_id",
      "required": true,
      "schema": {
       "title": "Decision Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/TimelineEntry"
         },
         "title": "Response Timeline Api V1 Decisions  Decision Id  Timeline Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Timeline",
    "tags": [
     "insights"
    ]
   }
  },
...
  "/api/v1/ingest/upload": {
   "post": {
    "operationId": "ingest_upload_api_v1_ingest_upload_post",
    "parameters": [
     {
      "in": "header",
      "name": "X-Request-ID",
      "required": true,
      "schema": {
       "title": "X-Request-Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": true,
      "schema": {
       "title": "Idempotency-Key",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Ingest Upload",
    "tags": [
     "insights"
    ]
   }
  },
...
  "/api/v1/models/{name}/promote": {
   "post": {
    "operationId": "model_promote_api_v1_models__name__promote_post",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "X-Request-ID",
      "required": true,
      "schema": {
       "title": "X-Request-Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": true,
      "schema": {
       "title": "Idempotency-Key",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Model Promote",
    "tags": [
     "insights"
    ]
   }
  },
  "/api/v1/models/{name}/rollback": {
   "post": {
    "operationId": "model_rollback_api_v1_models__name__rollback_post",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "X-Request-ID",
      "required": true,
      "schema": {
       "title": "X-Request-Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": true,
      "schema": {
       "title": "Idempotency-Key",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Model Rollback",
    "tags": [
     "insights"
    ]
   }
  },
  "/api/v1/models/{name}/{version}": {
   "get": {
    "operationId": "model_detail_api_v1_models__name___version__get",
    "parameters": [
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "version",
      "required": true,
      "schema": {
       "title": "Version",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ModelDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Model Detail",
    "tags": [
     "insights"
    ]
   }
  },
  "/api/v1/objective": {
   "get": {
    "operationId": "objective_api_v1_objective_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WorkspaceObjective"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Objective",
    "tags": [
     "insights"
    ]
   },
   "put": {
    "operationId": "objective_change_api_v1_objective_put",
    "parameters": [
     {
      "in": "header",
      "name": "X-Request-ID",
      "required": true,
      "schema": {
       "title": "X-Request-Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": true,
      "schema": {
       "title": "Idempotency-Key",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WorkspaceObjective"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Objective Change",
    "tags": [
     "insights"
    ]
   }
  },
...
  "/api/v1/policy": {
   "get": {
    "operationId": "policy_api_v1_policy_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PolicyOut"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Policy",
    "tags": [
     "insights"
    ]
   },
   "put": {
    "operationId": "policy_change_api_v1_policy_put",
    "parameters": [
     {
      "in": "header",
      "name": "X-Request-ID",
      "required": true,
      "schema": {
       "title": "X-Request-Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": true,
      "schema": {
       "title": "Idempotency-Key",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PolicyOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Policy Change",
    "tags": [
     "insights"
    ]
   }
  },
...
  "/api/v1/workspaces": {
   "get": {
    "operationId": "workspaces_api_v1_workspaces_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WorkspaceList"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Workspaces",
    "tags": [
     "insights"
    ]
   },
   "post": {
    "operationId": "workspace_create_api_v1_workspaces_post",
    "parameters": [
     {
      "in": "header",
      "name": "X-Request-ID",
      "required": true,
      "schema": {
       "title": "X-Request-Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": true,
      "schema": {
       "title": "Idempotency-Key",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Workspace"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Workspace Create",
    "tags": [
     "insights"
    ]
   }
  },
  "/api/v1/workspaces/{workspace_id}/activate": {
   "post": {
    "operationId": "workspace_activate_api_v1_workspaces__workspace_id__activate_post",
    "parameters": [
     {
      "in": "path",
      "name": "workspace_id",
      "required": true,
      "schema": {
       "title": "Workspace Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "X-Request-ID",
      "required": true,
      "schema": {
       "title": "X-Request-Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": true,
      "schema": {
       "title": "Idempotency-Key",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WorkspaceList"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Workspace Activate",
    "tags": [
     "insights"
    ]
   }
  }
 }
}